monitoring/requirements.txt
# DWH Monitoring Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
elasticsearch>=8.11.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
